update the solution for the second part
It is so much worse tha before.
One can change lists via lambda expressions and even concatenate multiple operations via "something if action1 or action2 or action3 else whatever". And because the lists variable is given as a default value at the creation of the dictionary, we can change its entries in the loop without  needing to pass or overwrite it explicitly later on.